// src/lib/components/data-picker/category-picker/data/category-detail-data.interface.ts
/**
 * @author ziyad.hajj-hassan
 */
export interface CategoryDetailDataInterface {
  plentyId: number;
  categoryId?: number;
  name?: string;
  canonicalLink?: string;
  description?: string;
  description2?: string;
  shortDescription?: string;
  metaDescription?: string;
  nameUrl?: string;
  previewUrl?: string;
  image?: string;
  image2?: string;
  itemListView?: string;
  singleItemView?: string;
  pageView?: string;
  placeholderTranslation?: string;
  fulltext?: string;
  lang?: string;
  metaKeywords?: string;
  metaTitle?: string;
  metaRobots?: string;
  position?: string;
  updatedAt?: string;
}